FB and login parts missing
